feat: enhance chat client processing with multi-turn support and tool handling
- Added maxTurns parameter to ChatClient's process method for controlling conversation length.
- Introduced OpenAIChatProcessor for shared processing logic, including tool call handling and message updates.
- Refactored OpenAIClient and OpenRouterClient to utilize OpenAIChatProcessor for improved message processing.
- Expanded test coverage for multi-turn conversations and UI tool interactions, ensuring robust functionality.
